chore: cleanup raw kernel client
